New Piercing System and Health Bar fix
Health bars just needed a blank texture. Oops. GameManager looks for MainCanvas tag now. New piercing system allows for projectiles to have a set amount of pierces, infinite piercing, and or repeat piercing (hitting the same enemy multiple times). Chain lightning is finally in an ok spot for once. Might need a bit of cleanup and sequential chaining for flair, but it's alright for now.