Mono Effects, Tooltips
Randomly added tooltips to places in the editor that would be helpful. Will add more as I go. Effects now have the option to only have one instance AND no stacking. Created script for a slow debuff, but didn't work on it much. For some reason Effect baseDuration is a rebellious little shit that refuses to serialize properly. This will be a problem for Tuesday me.